backend/requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
bcrypt==3.2.2
python-multipart
pydantic
requests
google-auth
httpx
boto3


